Create n-column-layout.html
This demo consists of 3 column layout. 
So child divs each of 33% width would be created. 
Additional catches of CSS :
float : left = enables the next child div's content to be included in the new line. So if there is no horizontal space left, the next div content will not reside on the same line.
overflow : hidden = hides the content exceeding the limit.